fix: refine provider-native tool discovery cleanup
