+++ packages/ai/src/constants/standalone-tool-discovery.ts
@@ -1,0 +1,9 @@
+export const DEFAULT_STANDALONE_TOOL_NAMES = [
+  "listAvailableSkills",
+  "getSkillByName",
+  "getAvailableIntegrations",
+  "webSearch",
+] as const;
+
+export const STANDALONE_TOOLING_DESCRIPTION =
+  "Notra app tools are available directly for this provider. Skills, integration discovery, Firecrawl, content, brand, GitHub, Linear, and post tools are included in the bounded built-in tool set.";
