+++ packages/ai/src/orchestration/orchestrate-standalone.ts
@@ -1,3 +1,7 @@
+import {
+  DEFAULT_STANDALONE_TOOL_NAMES,
+  STANDALONE_TOOLING_DESCRIPTION,
+} from "@notra/ai/constants/standalone-tool-discovery";
 import { getEnabledMcpServerCount } from "@notra/ai/integrations/mcp-tool-index";
 import { createModel } from "@notra/ai/model";
 import { getStandaloneChatPrompt } from "@notra/ai/prompts/standalone-chat";
@@ -27,10 +31,8 @@ import {
   stepCountIs,
   streamText,
   type Tool,
-  tool,
   type UIMessage,
 } from "ai";
-import { z } from "zod";
 import {
   hasEnabledGitHubIntegration,
   hasEnabledLinearIntegration,
@@ -46,28 +48,6 @@ import {
   getLinearContextFromIntegrations,
   getRepoContextFromIntegrations,
 } from "./standalone-tool-registry";
-
-const NOTRA_MANAGER_TOOL_NAMES = [
-  "searchNotraTools",
-  "activateNotraTools",
-  "listActiveNotraTools",
-  "deactivateNotraTools",
-] as const;
-
-const DEFAULT_STANDALONE_TOOL_NAMES = [
-  "listAvailableSkills",
-  "getSkillByName",
-  "getAvailableIntegrations",
-  "webSearch",
-] as const;
-
-const NOTRA_TOOLING_DESCRIPTION =
-  "Notra app tools are available through lazy discovery. Use searchNotraTools to find built-in content, brand, GitHub, Linear, and post tools by intent, then activateNotraTools before calling them. Basic skills, integration discovery, and web search tools are available by default when configured.";
-const WHITESPACE_REGEX = /\s+/;
-const LEGACY_NOTRA_TOOL_ALIASES: Record<string, string> = {
-  getBrandReferences: "getAvailableBrandReferences",
-  searchBrandReferences: "getAvailableBrandReferences",
-};
 
 export async function orchestrateStandaloneChat(
   input: StandaloneChatInput,
@@ -182,13 +162,10 @@ export async function orchestrateStandaloneChat(
     : null;
   const toolDiscoveryMode = providerNativeToolRuntime
     ? ("provider-native" as const)
-    : ("custom" as const);
-  const notraToolRuntime =
-    providerNativeToolRuntime ??
-    createStandaloneToolProvisioningRuntime({
-      tools: baseToolSet.tools,
-      defaultActiveToolNames,
-    });
+    : ("static" as const);
+  const notraToolRuntime = providerNativeToolRuntime ?? {
+    tools: baseToolSet.tools,
+  };
   const tools = providerNativeToolRuntime
     ? {
         ...providerNativeToolRuntime.tools,
@@ -209,7 +186,9 @@ export async function orchestrateStandaloneChat(
     : null;
 
   const descriptions = [
-    ...(providerNativeToolRuntime?.descriptions ?? [NOTRA_TOOLING_DESCRIPTION]),
+    ...(providerNativeToolRuntime?.descriptions ?? [
+      STANDALONE_TOOLING_DESCRIPTION,
+    ]),
     ...(providerNativeMcpRuntime?.descriptions ?? []),
     ...(lazyMcpRuntime?.descriptions ?? []),
   ];
@@ -366,128 +345,6 @@ function getInitialActiveToolNames({
   );
 }
 
-function createStandaloneToolProvisioningRuntime({
-  tools,
-  defaultActiveToolNames,
-}: {
-  tools: Record<string, Tool>;
-  defaultActiveToolNames: string[];
-}) {
-  const exposedTools: Record<string, Tool> = {};
-  const activeToolNames = new Set([
-    ...defaultActiveToolNames.filter((name) => name in tools),
-    ...NOTRA_MANAGER_TOOL_NAMES,
-  ]);
-  const managerToolNameSet = new Set<string>(NOTRA_MANAGER_TOOL_NAMES);
-  const defaultToolNameSet = new Set(defaultActiveToolNames);
-  const provisionableToolNames = Object.keys(tools).filter(
-    (name) => !(managerToolNameSet.has(name) || defaultToolNameSet.has(name))
-  );
-  const getActiveToolNames = () =>
-    Array.from(activeToolNames).filter(
-      (name) => name in exposedTools || managerToolNameSet.has(name)
-    );
-  for (const toolName of defaultActiveToolNames) {
-    if (toolName in tools) {
-      exposedTools[toolName] = tools[toolName] as Tool;
-    }
-  }
-  const managerTools: Record<string, Tool> = {
-    searchNotraTools: tool({
-      description:
-        "Search built-in Notra app tools by intent before activating them. Use this for content creation, post lookup, brand context, GitHub, Linear, and other Notra capabilities that are not currently active.",
-      inputSchema: z.object({
-        query: z.string().min(1),
-        limit: z.number().int().min(1).max(12).default(8),
-      }),
-      execute: async ({ query, limit }) => ({
-        results: searchProvisionableTools({
-          tools,
-          toolNames: provisionableToolNames,
-          query,
-          limit,
-          activeToolNames,
-        }),
-      }),
-    }),
-    activateNotraTools: tool({
-      description:
-        "Activate built-in Notra app tools for this chat run. Search first unless you already know the exact tool names.",
-      inputSchema: z.object({
-        toolNames: z.array(z.string().min(1)).min(1).max(8),
-        reason: z.string().max(500).optional(),
-      }),
-      execute: async ({ toolNames }) => {
-        const activated: Array<{
-          toolName: string;
-          description: string | undefined;
-        }> = [];
-        const unknown: string[] = [];
-        for (const requestedToolName of Array.from(new Set(toolNames))) {
-          const toolName = resolveNotraToolName(requestedToolName);
-          if (!(toolName in tools)) {
-            unknown.push(requestedToolName);
-            continue;
-          }
-          exposedTools[toolName] = tools[toolName] as Tool;
-          activeToolNames.add(toolName);
-          activated.push({
-            toolName,
-            description: tools[toolName]?.description,
-          });
-        }
-        return {
-          activated,
-          unknown,
-          activeTools: getActiveToolNames(),
-        };
-      },
-    }),
-    listActiveNotraTools: tool({
-      description: "List built-in Notra app tools currently active.",
-      inputSchema: z.object({}),
-      execute: async () => ({
-        activeTools: getActiveToolNames().filter(
-          (name) => !managerToolNameSet.has(name)
-        ),
-      }),
-    }),
-    deactivateNotraTools: tool({
-      description:
-        "Deactivate built-in Notra app tools that are no longer needed. Basic discovery tools and manager tools remain active.",
-      inputSchema: z.object({
-        toolNames: z.array(z.string().min(1)).min(1),
-      }),
-      execute: async ({ toolNames }) => {
-        const deactivated: string[] = [];
-        for (const requestedToolName of toolNames) {
-          const toolName = resolveNotraToolName(requestedToolName);
-          if (
-            defaultToolNameSet.has(toolName) ||
-            managerToolNameSet.has(toolName)
-          ) {
-            continue;
-          }
-          if (activeToolNames.delete(toolName)) {
-            delete exposedTools[toolName];
-            deactivated.push(toolName);
-          }
-        }
-        return {
-          deactivated,
-          activeTools: getActiveToolNames(),
-        };
-      },
-    }),
-  };
-  Object.assign(exposedTools, managerTools);
-
-  return {
-    tools: exposedTools,
-    getActiveToolNames,
-  };
-}
-
 function getDefaultStandaloneActiveToolNames({
   tools,
   context,
@@ -524,67 +381,6 @@ function getDefaultStandaloneActiveToolNames({
   }
 
   return Array.from(active);
-}
-
-function searchProvisionableTools({
-  tools,
-  toolNames,
-  query,
-  limit,
-  activeToolNames,
-}: {
-  tools: Record<string, Tool>;
-  toolNames: string[];
-  query: string;
-  limit: number;
-  activeToolNames: Set<string>;
-}) {
-  const terms = query
-    .toLowerCase()
-    .split(WHITESPACE_REGEX)
-    .map((term) => term.trim())
-    .filter(Boolean);
-
-  return toolNames
-    .map((toolName) => {
-      const description = tools[toolName]?.description ?? "";
-      const aliases = getLegacyAliasesForToolName(toolName);
-      const haystack =
-        `${toolName} ${aliases.join(" ")} ${description}`.toLowerCase();
-      const score = terms.reduce((total, term) => {
-        if (
-          toolName.toLowerCase().includes(term) ||
-          aliases.some((alias) => alias.toLowerCase().includes(term))
-        ) {
-          return total + 4;
-        }
-        if (haystack.includes(term)) {
-          return total + 1;
-        }
-        return total;
-      }, 0);
-      return {
-        toolName,
-        aliases,
-        description,
-        alreadyActive: activeToolNames.has(toolName),
-        score,
-      };
-    })
-    .filter((result) => result.score > 0 || terms.length === 0)
-    .sort((a, b) => b.score - a.score || a.toolName.localeCompare(b.toolName))
-    .slice(0, limit)
-    .map(({ score: _score, ...result }) => result);
-}
-
-function resolveNotraToolName(toolName: string) {
-  return LEGACY_NOTRA_TOOL_ALIASES[toolName] ?? toolName;
-}
-
-function getLegacyAliasesForToolName(toolName: string) {
-  return Object.entries(LEGACY_NOTRA_TOOL_ALIASES)
-    .filter(([, currentToolName]) => currentToolName === toolName)
-    .map(([legacyToolName]) => legacyToolName);
 }
 
 function isGitHubToolName(toolName: string) {
